Added authenticate for get and post todos

+++ server/server.js
@@ -63,9 +63,10 @@ app.delete('/users/me/token', authenticate, (req, res) => {
 	);
 });
 
-app.post('/todos', (req, res) => {
+app.post('/todos', authenticate, (req, res) => {
 	var todo = new Todo({
-		text: req.body.text
+		text: req.body.text,
+		_creator: req.user._id
 	});
 
 	todo.save().then(
@@ -78,8 +79,8 @@ app.post('/todos', (req, res) => {
 	);
 });
 
-app.get('/todos', (req, res) => {
-	Todo.find().then(
+app.get('/todos', authenticate, (req, res) => {
+	Todo.find({ _creator: req.user._id }).then(
 		todos => {
 			res.send({ todos });
 		},
@@ -89,14 +90,17 @@ app.get('/todos', (req, res) => {
 	);
 });
 
-app.get('/todos/:id', (req, res) => {
+app.get('/todos/:id', authenticate, (req, res) => {
 	var id = req.params.id;
 
 	if (!ObjectID.isValid(id)) {
 		return res.status(404).send();
 	}
 
-	Todo.findById(id)
+	Todo.findOne({
+		_id: id,
+		_creator: req.user._id
+	})
 		.then(todo => {
 			if (!todo) {
 				return res.status(404).send();
@@ -109,7 +113,7 @@ app.get('/todos/:id', (req, res) => {
 		});
 });
 
-app.delete('/todos/:id', (req, res) => {
+app.delete('/todos/:id', authenticate, (req, res) => {
 	var id = req.params.id;
 
 	// console.log(id)
@@ -119,7 +123,7 @@ app.delete('/todos/:id', (req, res) => {
 	}
 
 	// Todo.findByIdAndRemove(id)
-	Todo.findOneAndDelete({ _id: id })
+	Todo.findOneAndDelete({ _id: id, _creator:req.user._id })
 		.then(todo => {
 			if (!todo) {
 				return res.status(404).send();
